fix: unify win loss flat metrics (closes #258)
